requirements.txt
streamlit>=1.31.0
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
plotly>=5.18.0
altair>=5.2.0
sqlite-utils>=3.36
tenacity>=8.2.0
cachetools>=5.3.0
mcp[cli]>=1.0.0
yfinance>=0.2.36
pytest>=8.0.0
pytest-asyncio>=0.23.0
